mata: Add patched uceShimService.apk
This fixes a rescue party on SIM hotswap (and maybe other misc. bugs).

The root cause of the issue is RCSServiceStub calls get(0) on the
list of SubscriptionInfo objects returned from SubscriptionManager.getActiveSubscriptionInfoList().
RCSServiceStub checks if the list is null and aborts and then checks
if the size is greater than 1 to determine if the device is MSIM.
Unfortunately, it doesn't check to see if the list is empty. This causes
a crash which eventually triggers rescue party on a device.

In order to fix this, the offending method (getDefaultIccId()) is edited
to also abort in the case that the SubscriptionInfo list is empty.

The change in jadx output is pasted below:

diff --git a/sources/com/qualcomm/qti/uceShimService/RCSServiceStub.java b/sources/com/qualcomm/qti/uceShimService/RCSServiceStub.java
index d1b3397..a7e3bc2 100644

--- a/sources/com/qualcomm/qti/uceShimService/RCSServiceStub.java
+++ b/sources/com/qualcomm/qti/uceShimService/RCSServiceStub.java
@@ -51,12 +51,13 @@ public class RCSServiceStub extends UceServiceBase {
     }

     private boolean getDefaultIccId() {
+        int size;
         SubscriptionManager subMgr = SubscriptionManager.from(this.servContext);
         List<SubscriptionInfo> subInfoList = subMgr.getActiveSubscriptionInfoList();
-        if (subInfoList == null) {
+        if (subInfoList == null || (size = subInfoList.size()) == 0) {
             return false;
         }
-        if (subInfoList.size() > 1) {
+        if (size > 1) {
             Log.i(logTAG, "RCSStartRCSService Detected MultiSim");
         }
         this.mIccId = subInfoList.get(0).getIccId();

The useShimService.apk from the QP1A.190711.148 vendor release doesn't
have this issue, but that requires additional methods in com.qualcomm.qti.uceservice@1.0::IOptionsListener
that causes issues with loading it. Other than those additional methods,
this is the only real change in the QP1A.190711.148 APK.

Change-Id: I1a5cb6a024ed2ed8a4a9e5e0dc9d162b49ec8967